requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
llama-cpp-python==0.2.79
pydantic==2.5.3
python-dotenv==1.0.0
pydantic-settings==2.1.0
aiofiles==23.2.1
httpx==0.25.2
numpy==1.24.3
